Began the implementation of named observer functions for SSTs.
Implementation is not yet complete, since constructing an SST with named
functions will not do anything useful, but the types and structures for
declaring named functions are implemented, as is the template machinery
for type-checking them when they are registered with an SST. The new SST
constructor degrades gracefully to the old one when no named functions
are provided.

Unfortunately, in order to implement this change without adding template
bloat to the Predicates class, SST's predicates member must be changed
to a pointer rather than a value type. This seems unavoidable for now.