Eliminated deadlock by releasing lock during trigger invocations
Holding a lock on predicates_mutex while calling a trigger function
could cause deadlock if the trigger itself tried to lock a different
mutex. Now we give up the lock on predicates_mutex before calling the
trigger and re-aquire it upon return, and take precautions to ensure
that nothing the trigger does can invalidate the iterator over the
predicates list (which is what predicates_mutex protects). This required
replacing the "list of triggers" with a single shared_ptr to a trigger,
but users can still register a list of triggers with a single predicate
(the insert function internally wraps this list in a lambda that calls
each function in sequence).